(#20) Split keycode.c
