+++ src/win32/keycode.c
@@ -1,0 +1,6 @@
+#include "../keycode.h"
+
+MMKeyCode keyCodeForChar(const char c)
+{
+	return VkKeyScan(c);
+}
